feat: complete Phase A–V11 — NURBS kernel, 13 primitives, 10 workbenches, FreeCAD-level UI
- NURBS kernel: basis functions, fitting, trimmed surfaces, SSI marching (A01–A28)
- Geometry binding: exact B-Rep with face/edge surface/curve binding (B01–B14)
- STEP/IGES/DXF/PLY/3MF/BREP import-export (10 file formats)
- Features: extrude, revolve, fillet, chamfer, draft, sweep, loft, pad, pocket, hole
- Booleans: union, subtract, intersect, XOR with face-splitting preprocessing
- 13 primitives, 24 sketch constraints, 6 sketch entities, assembly solver
- Draft/Surface/FEM/TechDraw workbenches with full backend
- Viewer UI: split gui.rs into 11 submodules, hierarchical model tree,
  property editor, report panel, context menus, enhanced status bar
- 739 tests, zero warnings

+++ crates/topology/src/handle.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 #[serde(bound = "")]
 pub struct Handle<T> {
     pub(crate) index: u32,
-    pub(crate) generation: u32,
+    pub(crate) generation: u64,
     #[serde(skip)]
     pub(crate) _marker: PhantomData<T>,
 }
@@ -45,7 +45,7 @@ impl<T> Handle<T> {
 
     /// Returns the generation counter used for dangling-handle detection.
     #[inline]
-    pub fn generation(&self) -> u32 {
+    pub fn generation(&self) -> u64 {
         self.generation
     }
 
@@ -58,7 +58,7 @@ impl<T> Handle<T> {
     /// or fabricated values will not cause UB but will make subsequent lookups
     /// return `None`.
     #[inline]
-    pub fn from_raw_parts(index: u32, generation: u32) -> Self {
+    pub fn from_raw_parts(index: u32, generation: u64) -> Self {
         Self {
             index,
             generation,
